Added support for arithmetic with different types
Adds support for operations like int16_t + int64_t or
uint8_t + double(the operand with the smaller size is promoted to
the operand with the larger size) and tests for this new functionality.
Also adds a few helper functions including StaticCastOrConvertLiteral,
CastOperandAndDoArithmetic, and CastOperandAndDoComparison.
Also adds a GetAs<T> member to the AstLiteralExpression to get the
value of a literal as type T